requirements.txt
accelerate==1.13.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorlog==6.10.1
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
docling==2.82.0
docling-core==2.71.0
docling-ibm-models==3.13.0
docling-parse==5.6.2
docstring_parser==0.17.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Faker==40.11.1
filelock==3.25.2
filetype==1.2.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
latex2mathml==3.79.0
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.19
narwhals==2.18.1
networkx==3.4.2
numpy==2.2.6
ocrmac==1.0.1
omegaconf==2.3.0
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
polyfactory==3.3.0
portalocker==3.2.0
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pylatexenc==2.10
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-pptx==1.0.2
pytz==2026.1.post1
PyYAML==6.0.3
qdrant-client==1.17.1
rapidocr==3.7.0
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semchunk==3.2.5
sentence-transformers==5.3.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.55.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xlsxwriter==3.2.9
